weston: add configuration option for no input device
With this patch, it is now possible to override the required
input devices at launch.

In weston.ini,
  [libinput]
  require_input=true
is the default behavior, and will make Weston to bail out when
no input device is found.

With
  [libinput]
  require_input=false
it may be possible to still run Weston without any input
device. This patch updates weston.ini to this setting.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>