weston-init: Update systemd service unit.
Based on the "ultimate" systemd service unit [1], this unit
works without the need of weston-start/weston-launch, as is
recommended by Weston developers.

It handles:
* proper VT/tty allocation and handling,
* PAM login (XDG environment settings),
* correct logging,
* optional arguments to Weston via config file.

N.B.: We rely on udev to make a request for this service as
soon as the DRM card is found. Therefore, while the following
line makes a lot of sense, it needs to be commented out so
that the service is not immediately launched opon boot:
  #Alias=multi-user.target.wants/weston.service

[1] https://phabricator.freedesktop.org/T63

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>