weston-init: initiate PAM session for user root
Some time ago (636156d), weston-launch used to set up the
session for user root as well, but this behavior was
changed to only set PAM up (and thus login environment,
such as the XDG variables) when instructed to be launched
as another user. This other user, however, can be root, and
all will be well.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>