machine: mount root device read-write on boot
systemd recipe from oe-core is creating an empty machine-id file.
If we want to change this behavior and remove the file, it triggers the
following issue:
systemd[1]: System cannot boot: Missing /etc/machine-id and /etc is mounted read-only.
systemd[1]: Booting up is supported only when:
systemd[1]: 1) /etc/machine-id exists and is populated.
systemd[1]: 2) /etc/machine-id exists and is empty.
systemd[1]: 3) /etc/machine-id is missing and /etc is writable

Mount the root device read-write to allow systemd to populate /etc.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>