edk2-hikey960: add recipe to build UEFI for HiKey960
Until it's upstreamed, we need to use a fork of UEFI tools which include
HiKey960 platform configuration. As a result, adjust other EDK2 recipes so
we can override UEFI tools in the board specific recipe.

For HiKey960, we don't use a separate recipe for l-loader. It's built as
part of the edk2 recipe and doesn't depends on an aarch32 toolchain

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>