linux-hikey960: Replace branch/rebaseable with nobranch
Currently this recipe will look for a specific SHA-1 hash on the (rebaseable)
hikey960-upstream-rebase branch, however even with the SHA-1 hash preserved
using a tag, we are not able to download it once the branch has been rebased.

Since the SHA-1 *is* preserved usinga tag, we can fix this by enabling
nobranch, we also remove branch for being redundant and rebaseable for being
pointless (it does little more than presents user with a warning).

Signed-off-by: Daniel Thompson <daniel.thompson@linaro.org>