linux.inc: remove firmware_install
firmware_install target has been removed from mainline kernel in commit
5620a0d1aacd554ebebcff373e31107bb1ef7769

[...]
The last firmware change for the in-kernel firmware source code was back
in 2013.  Everyone has been relying on the out-of-tree linux-firmware
package for a long long time.

So let's drop it, it's baggage we don't need to keep dragging around
(and having to fix random kbuild issues over time...)
[...]

Remove also capes copy since it isn't generic and doesn't apply to
mainline kernels. If we need capes, we can handle it outside of linux.inc.

Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>