atf-juno: move the ramdisk up in NOR flash to give more space for larger kernel
Move the ramdisk up in NOR flash to give more space for a larger kernel.
This also means that we have less space for the ramdisk, however, on OE
systems, we use a stub ramdisk of 576 bytes, so we don't need much space.
This will probably break Android, which uses a 1.5MB ramdisk.

Signed-off-by: Ryan Harkin <ryan.harkin@linaro.org>
Signed-off-by: Fathi Boudra <fathi.boudra@linaro.org>