gst-plugins-bad: establish build/run dependencies against OpenGL
When building with Mali450 (instead of Mesa or a different
OpenGL provider), some libraries would have been reported
as missing during the package_qa phase.

This change establishes the build and run dependencies so
that the proper OpenGL implementation is pulled.

The recipe upon which this is appended, includes this line:
  PACKAGECONFIG[egl]             = "--enable-egl,--disable-egl,virtual/egl"
That is changed here to:
  PACKAGECONFIG[egl]             = "--enable-egl,--disable-egl,virtual/egl,libegl"
so that the run-time dependency is added as well.

The same is done for gles2 and opengl.

Packages libgstgl-1.0 and gstreamer1.0-plugins-bad-opengl
have their runtime dependencies added directly.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>