mali450-userland: fix SONAME for libMali.so
The ELF header of libMali does not include the SONAME and
that creates a tremendous amount of issues with package_qa
as a lot of binaries cannot be matched against their r/build
dependency.

We will be forever grateful to Khem Raj for his insight on
this matter. We heart you.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>