grub: drop custom recipe and migrate to the upstream grub-efi version
Instead of duplicating an entire recipe which is now older than what is
provided via OE-core, drop the current recipe (grub-git) in favor of the
upstream grub-efi (2.02) version.

Upstream is mostly in sync with grub-git, only missing piece is the grub
cfg customization, which is currently replaced via an extra grub-mkimage
call.

Signed-off-by: Ricardo Salveti <ricardo@opensourcefoundries.com>