linux-generic: direct branch instr modules need PLT for relocation
Fixes: https://bugs.linaro.org/show_bug.cgi?id=3769

ARM branch (BL) instructions can't branch to more than 32 MB from the
BL instruction. Unfortunately, this is not enough for arm32 (and,
sometimes, for arm64), causing symbol relocation issues when trying to
load kernel modules (depending on base address for the modules, amount
of already loaded modules and the delta in between the branch
instruction and relocation target: a kernel symbol in this case).

Enabling:
    arm32: CONFIG_ARM_MODULE_PLTS
    arm64: CONFIG_ARM64_MODULE_PLTS

will make kernel linker code, when loading the kernel module, to
execute those branches as some form of indirect branch (using PLTs).
Without this, there would be a need for ALL kernel modules to be built
with: "-mlong-calls", making all branches indirect (and affecting
performance). With this, there is a slight increase in size in the
generated kernel objects (5% according to commit: 7d485f647c), but the
direct branches are still used in other places.

Signed-off-by: Rafael David Tinoco <rafael.tinoco@linaro.org>