linux stable-rc/mainline/next/generic: enable KSM
As per the documentation, Kernel Samepage Merging is a
memory-saving de-duplication feature, enabled by CONFIG_KSM=y
and activated via sysfs.

When enabled in the kernel, the default is to not do anything
at all, until it is activated with:
  echo 1 > /sys/kernel/mm/ksm/run

Available since 2.6.32. More information here:
  https://www.kernel.org/doc/Documentation/vm/ksm.txt

This change here is to ensure that the same configuration is
effectively used across all architectures, as noted in bug
report #3857:
  https://bugs.linaro.org/show_bug.cgi?id=3857

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>