linux-generic: Remove qcs404 from deploy on mainline, next, generic
As of next-20181119, a new Qualcomm board has been added to
Linux, which seems problematic when dealing with dtbTool. The
workaround has been to remove those files before they present
any problems, so this is in line with the current behavior.

Currently, only linux-next is affected, but it will eventually
reach mainline, and might bit if using the generic kernel as
that can build any revision whatsoever.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>