linux: Remove LKFT recipes for stable-rc, mainline and next
These recipes live now in meta-lkft and are not needed by RPB.

The plan is to move as much as possible off of meta-96boards,
as having two copies of the same files is a straight path to
confusion and frustration. Frustration leads to anger. Anger
leads to hate. Hate leads to suffering.

Signed-off-by: Daniel Díaz <daniel.diaz@linaro.org>

(Cherry-picked from commit 2ce67ee364bc984569716c833df9c3c974d6f392)